New typeconverter for converting html form data.
Right now all it converts is bools from on/off/null, Since everything
else seems to be covered by regular type conversion.

This allows us to have a bool for a checkbox. I'm not 100% sure every
browser sends 'on' for a checkbox, but if it does that will get
converted to true. null will get converted to false.

"true"/"false" will be handled by the regular converter.