All the seeking/writing tests are now passing. Which just proves I need more tests.
Fixup seeking back into a stream and writing over data. This seems to be working
now, but probably needs more testing. Particularly verifying the data (since most
of the tests just verify lenth/position) and tests around the edges, like writing
to the exact end spot on a buffer.

All stubbed in the Insert method.  This will be used to prepend things like
HTTP Headers to the stream.