Switch to the new HttpResponseStream for output.
Using a proper System.IO.Stream will let people using things like
StreamWriters, XmlReaders, ect on the output. Also they'll be able
to seek to the beginning of the stream and read the stream using
normal stream readers.

This should make things like HttpPipes that load and morph the
output html using the html agility pack easier.  An obvious use-case
would be finding the <form> element in an html output and adding
a csrf token.