Start work on moving the write queue into the IOStream.
The write queue is going to be moved into the IOStream so it can be
removed from the HttpTransaction and allow the HttpResponseStream to
deal directly with the IOStream. This decouples things a little and
is much cleaner.

Part of the goal here is to remove buffering and use
chunked encoding by default.  If people want to set Content-Length
they can do it manually (or maybe via some config/IOC mechanism).

Also allow multiple write byte operations to be combined into a single
op.  This means we might be able to scatter/gather them all at once.