Go back to a separate read/write watcher.
If we don't disable writing until data is ready, we will max out the CPU on
long requests that are asyncronously waiting for data to be queued up.  ie nothing
is in the write_ops queue yet, but the browser is ready for data.

Unfortunately I can't get ev_io_set to work for enable/disable, so for now we'll
just use two watchers. Hopefully in the future I can get setting to work.