Use lib/manos/ instead of build for output.
Rationale:
If running from inside MD, rather than from a local install, we cannot
access the files like layout, and other static content needed for error
and health monitors.

Making the output in this format enables us to easily symlink data as ./share/manos
and the behaviour will be identical whether Manos is run as an installed app
or it's run as a local debugging version.

tl;dr - it'll make hacking on more advanced features easier while not touching
the existing codebase, and the workflow for current hackers of Manos