IOStream is now a base class and SocketStream replaces it.
Eventually there will be two main IOStreams.  SocketStream and FileStream
this move is mainly a rename.  Previously IOStream was used where the
stream should only be a Socket.

In the future I might relax this a bit and use the base class (or an
interface) in more places to ease testing.  For now I'd like to keep
too much abstraction out of the IO layer.