abstract base class for async streams
writes can be handled as IEnumerable<ByteBuffer> just as effectively as
with IWriteOperations. sequences of buffers remove the ability for other
write ops, such as SendFile, but other operations can be implemented on
top of normal, sequenced writes, or as special interfaces for particular
instances of streams.
that way, we have a cleaner stream interface that also allows filter
hierarchies without much effort.