coerce managed io to work correctly with the current http handling system
sadly, this has got to be one of the most awful hacks in existence. we
now do not handle transaction end code after the transaction has (from a
user pov) ended, but after everything was written by the stream. this is
achieved by writing an empty stream to the socket that invokes a
delegate which then posts a oneshot watcher to the socket's context,
which finally closes the socket and handles the other transaction end
stuff.

i am not proud.