Updates Environment.cs to have the correct paths on Windows, updates DocsModule.cs to have the tutorials work the same way as the manuals and to use working routing rules.
This updates the windows Environment so the directories work how the docs advertise, with all binaries in the prefix folder, docs in a "docs" subfolder, and layouts in a "layouts" subfolder. This also gets the docs feature in manostool working by having the tutorials be mapped by name to the correct file (like manuals are) and using the right routing rules.